Move topological sorting algorithm to separate class
