+++ src/DependencyGraph/Graph.ts
@@ -5,24 +5,9 @@
 
 import {SimpleCellAddress} from '../Cell'
 import {SimpleCellRange} from '../AbsoluteCellRange'
+import {TopSort, TopSortResult} from './TopSort'
 
 export type DependencyQuery<T> = (vertex: T) => [(SimpleCellAddress | SimpleCellRange), T][]
-
-export interface TopSortResult<T> {
-  sorted: T[],
-  cycled: T[],
-}
-
-// interface GraphNode<T> {
-//   id: number,
-//   data: T,
-// }
-
-enum NodeVisitStatus {
-  ON_STACK,
-  PROCESSED,
-  POPPED,
-}
 
 /**
  * Provides graph directed structure
@@ -33,9 +18,6 @@ enum NodeVisitStatus {
  * - this.edges(node) is subset of this.nodes (i.e. it does not contain nodes not present in graph) -- this invariant DOES NOT HOLD right now
  */
 export class Graph<T> {
-  // private _nodes: Set<GraphNode<T>> = new Set()
-  // private nextId = 0
-
   public nodes: Set<T> = new Set()
   public specialNodes: Set<T> = new Set()
   public specialNodesStructuralChanges: Set<T> = new Set()
@@ -202,112 +184,8 @@ export class Graph<T> {
    * @param onCycle - action to be performed when node is on cycle
    */
   public getTopSortedWithSccSubgraphFrom(modifiedNodes: T[], operatingFunction: (node: T) => boolean, onCycle: (node: T) => void): TopSortResult<T> {
-
-    const entranceTime: Map<T, number> = new Map()
-    const low: Map<T, number> = new Map()
-    const parent: Map<T, T> = new Map()
-    const inSCC: Set<T> = new Set() // do we need these data structures? logarithmic factor?
-
-    // node status life cycle:
-    // undefined -> ON_STACK -> PROCESSED -> POPPED
-    const nodeStatus: Map<T, NodeVisitStatus> = new Map()
-    const order: T[] = []
-
-    let time: number = 0
-
-    const sccNonSingletons: Set<T> = new Set()
-
-    modifiedNodes.reverse()
-    modifiedNodes.forEach((v: T) => {
-      if (nodeStatus.get(v) !== undefined) {
-        return
-      }
-      const DFSstack: T[] = [v]
-      const SCCstack: T[] = []
-      nodeStatus.set(v, NodeVisitStatus.ON_STACK)
-      while (DFSstack.length > 0) {
-        const u = DFSstack[DFSstack.length - 1]
-        // eslint-disable-next-line @typescript-eslint/no-non-null-assertion
-        switch (nodeStatus.get(u)!) {
-          case NodeVisitStatus.ON_STACK: {
-            entranceTime.set(u, time)
-            low.set(u, time)
-            SCCstack.push(u)
-            time++
-            this.adjacentNodes(u).forEach((t: T) => {
-              if (entranceTime.get(t) === undefined) {
-                DFSstack.push(t)
-                parent.set(t, u)
-                nodeStatus.set(t, NodeVisitStatus.ON_STACK)
-              }
-            })
-            nodeStatus.set(u, NodeVisitStatus.PROCESSED)
-            break
-          }
-          case NodeVisitStatus.PROCESSED: { // leaving this DFS subtree
-            let uLow: number
-            // eslint-disable-next-line @typescript-eslint/no-non-null-assertion
-            uLow = entranceTime.get(u)!
-            this.adjacentNodes(u).forEach((t: T) => { // Ta petla chyba jest niepotrzebna. Chiecko mogloby updatowac low[parent]
-              if (!inSCC.has(t)) {
-                if (parent.get(t) === u) {
-                  // eslint-disable-next-line @typescript-eslint/no-non-null-assertion
-                  uLow = Math.min(uLow, low.get(t)!)
-                } else {
-                  // eslint-disable-next-line @typescript-eslint/no-non-null-assertion
-                  uLow = Math.min(uLow, entranceTime.get(t)!)
-                }
-              }
-            })
-            low.set(u, uLow)
-            if (uLow === entranceTime.get(u)) {
-              const currentSCC: T[] = []
-              do {
-                currentSCC.push(SCCstack[SCCstack.length - 1])
-                SCCstack.pop()
-              } while (currentSCC[currentSCC.length - 1] !== u)
-              currentSCC.forEach((t) => {
-                inSCC.add(t)
-              })
-              order.push(...currentSCC)
-              if (currentSCC.length > 1) {
-                currentSCC.forEach((t) => {
-                  sccNonSingletons.add(t)
-                })
-              }
-            }
-            DFSstack.pop()
-            nodeStatus.set(u, NodeVisitStatus.POPPED)
-            break
-          }
-          case NodeVisitStatus.POPPED: { // it's a 'shadow' copy, we already processed this vertex and can ignore it
-            DFSstack.pop()
-            break
-          }
-        }
-      }
-    })
-
-    const shouldBeUpdatedMapping = new Set(modifiedNodes)
-
-    const sorted: T[] = []
-    const cycled: T[] = []
-    order.reverse()
-    order.forEach((t: T) => {
-      const adjacentNodes = this.adjacentNodes(t)
-      // eslint-disable-next-line @typescript-eslint/no-non-null-assertion
-      if (sccNonSingletons.has(t) || adjacentNodes.has(t)) {
-        cycled.push(t)
-        onCycle(t)
-        adjacentNodes.forEach((s: T) => shouldBeUpdatedMapping.add(s))
-      } else {
-        sorted.push(t)
-        if (shouldBeUpdatedMapping.has(t) && operatingFunction(t)) {
-          adjacentNodes.forEach((s: T) => shouldBeUpdatedMapping.add(s))
-        }
-      }
-    })
-    return {sorted, cycled}
+    const topSortAlgorithm = new TopSort<T>(this.edges)
+    return topSortAlgorithm.getTopSortedWithSccSubgraphFrom(modifiedNodes, operatingFunction, onCycle)
   }
 
   public getDependencies(vertex: T): T[] {
